Security Update
Store the secret into the application once it performed a successful QR
Code read so it could the application could refer to these secret if the
user want to fix the deivce without scanning a QR Code.
Previously, the application read the secret from the settings stored in
the controller, which means the user could easily modify these settings,
which makes it less secured.